Adams-call/requirements.txt
# WebRTC Communication App Dependencies
aiortc>=1.6.0
aiohttp>=3.9.0
aiohttp-session>=2.12.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
aiofiles>=23.2.0
